refactor: separate HTTP app from process lifecycle
